+++ funpayparsers/types/finances.py
@@ -36,6 +36,19 @@ class TransactionPreview(FunPayObject):
     withdrawal_number: str | None
     """Withdrawal card / phone / wallet number, if applicable."""
 
+    @property
+    def timestamp(self) -> int:
+        """
+        Transaction timestamp.
+
+        ``0``, if an error occurred while parsing.
+        """
+        from funpayparsers.parsers.utils import parse_date_string
+        try:
+            return parse_date_string(self.date_text)
+        except ValueError:
+            return 0
+
 
 @dataclass
 class TransactionInfo(FunPayObject):